Further code reorganization and bug fixes
I resolved some functionality problems when resizing elements - you can
now drag elements while resizing them, and the correct styles should be
added to the stylesheet. File exporting should also preserve the user
css sheet (changed innerHTML to textContent seems to have resolved the
issue). I also moved the modules closer to the top of the file so I
could reorganize controller functions to be closer to the top. I also
renamed controller functions to make generating the menu easier to
accomplish when it is represented by an object near the top of the file
and converted to views at runtime. Changes are synced between dev and
live files.