Incorporating Rangyinputs, bug fixes
I've implemented some automatic code formatting into the User CSS window
using the rangyinputs jQuery plug-in. The current space indentation for
CSS rules is 4 spaces, and this will be editable in the future once I
incorporate the storedObject module with a User Preferences menu item. I
also revised the stylesheet module so that CSS rule properties are not
reordered when dragging elements. The resize interaction now uses
"start" and "stop" functions similar to the resize actions, and I tried
to preserve allowing resizing and dragging to be active on the same
element.