Replaced Rangyinputs with Emmet textarea
I am not sure if I will be able to change the Emmet behavior of adding
tabs before CSS rules instead of spaces, which is why I only commented
out my rangyinputs key handler functionality. I updated the prompt
window CSS to allow expanding the textarea's size while preserving its
centering. I updated the Prompt module to account for providing both
strings and jQuery objects as the window's content, which should allow
it to avoid adding and removing elements from the DOM each time the
prompt window is open (this is said to cause memory issues on
single-page apps). I also removed the stopPropagation keyhandler for the
prompt window so Emmet could interact with the Insert HTML textarea
element. Emmet listens for evens on the document body, which means none
of the app elements can stop propagation if they need to access Emmet's
functionality. I also moved the views for the three insert menu items to
the main views object. The stylesheet module was updated with changes
made in the last commit.