Folder restructuring and fixes for Save and Restore
I'm still trying to find the best workflow to create a good development
environment that is easy to push to production. Today a friend of mine
helped me move Remarklet to DigitalOcean so if you are loading it from
remarklet.com it should be coming through minified and gzipped!
Unfortunately the secure certificate hasn't switched over yet so I
switched the bookmarklet links to use http. Sorry if this caused any
frustration :(