Further changes to file management, some upgrades
I changed my modules to work with RequireJS module definition
guidelines. I'm testing what kind of perf impact separating the files
has on load time. I also reduced the depth of the folder structure on
the server. I've moved the files over to a DigitalOcean server and am
trying to learn Sublime with SFTP and a bunch of other plug-ins.