Embed local image using Insert > Image
HTML5 has a really simple way to use the file input element to convert a
local image to a Data URI, and since this allows the image to be
embedded and saved with the exported page I've implemented this solution
in favor of image blobs. I also added some styles for the menu to keep
it from looking different on other sites, and there are other elements I
need to do this with. I may have forgot to mention in the last commit
that I solved some bugs with syncing CSS changes between page elements
and the CSS Changes view when they are being nudged.