Mostly Visual Changes, Some w/ CSS Editor & Shortcut
I've manually set some CSS properties in the UI that were being
overridden on some of the sites I tested it with. I also changed the CSS
selector hint in the CSS editor to be more helpful, and moved it below
the editor for more space. I also changed the "Finish Resizing" keyboard
shortcut to "Ctrl + Enter" to match the "Finish Editing Text" shortcut.
I also restored HTTPS to the main site so this bookmarklet will once
again load on secure pages which do not disallow it due to a content
security policy.