Removed Google Analytics
Back when I started the project, it wasn't MIT licensed, and I wanted to
know if anyone was finding it and using it on their own. Now it is MIT
licensed and on Github, and I no longer feel comfortable loading GA with
the script. I'm also planning to re-implement emmet's old textarea
plug-in to add more functionality to the CSS editor and the Insert HTML
prompt.