Using SASS, Fixed Resizing Functionality
I have established a workflow that allows me to build SASS files in
Sublime Text, so I restructured some of the selectors in the process and
will import a modified version of the jQuery UI CSS file. I also
upgraded jquery-ui.js to the latest version while keeping the same
plugins for it. I fixed an issue with the keyboard shortcut for ending
resizing that did not complete the cycle due to styles being removed
from the element prematurely.